refactor [detectors,fock] - use heterodyne sampling on fock module
- deletes the unused module `homodyne.py`
- implements all heterodyne sampling functions in `fock.py`
- refactors detector to use the fock module
